+++ src/models/scaleway_period_baremetal_period_v1_period_memory.rs
@@ -21,7 +21,7 @@ pub struct ScalewayPeriodBaremetalPeriodV1PeriodMemory {
     pub r#type: Option<String>,
     /// Frequency of the memory in MHz.
     #[serde(rename = "frequency", skip_serializing_if = "Option::is_none")]
-    pub frequency: Option<i32>,
+    pub frequency: Option<u64>,
     /// True if the memory is an error-correcting code memory.
     #[serde(rename = "is_ecc", skip_serializing_if = "Option::is_none")]
     pub is_ecc: Option<bool>,
