+++ src/models/scaleway_period_rdb_period_v1_period_node_type.rs
@@ -24,7 +24,7 @@ pub struct ScalewayPeriodRdbPeriodV1PeriodNodeType {
     pub description: Option<String>,
     /// Number of virtual CPUs.
     #[serde(rename = "vcpus", skip_serializing_if = "Option::is_none")]
-    pub vcpus: Option<i32>,
+    pub vcpus: Option<u64>,
     /// Quantity of RAM. (in bytes)
     #[serde(rename = "memory", skip_serializing_if = "Option::is_none")]
     pub memory: Option<u64>,
